rename MainView to WeatherResultView
